adding ui for add/import acct screens plus consolidating components
